Fin project cards. More work can be done
Currently my solution for the orange line on the side of the card
involves absolutely positioning a span element with orange color, I
would like to see if i can give an inset box-shadow on the card and make
it work.
Separately, i would like to implement a dynamic grid column structure.